Prepare linux-4.9 for h6

+++ drivers/pinctrl/intel/pinctrl-cherryview.c
@@ -1060,7 +1060,7 @@ static int chv_config_get(struct pinctrl_dev *pctldev, unsigned pin,
 }
 
 static int chv_config_set_pull(struct chv_pinctrl *pctrl, unsigned pin,
-			       enum pin_config_param param, u32 arg)
+			       enum pin_config_param param, u16 arg)
 {
 	void __iomem *reg = chv_padreg(pctrl, pin, CHV_PADCTRL0);
 	unsigned long flags;
@@ -1152,7 +1152,7 @@ static int chv_config_set(struct pinctrl_dev *pctldev, unsigned pin,
 	struct chv_pinctrl *pctrl = pinctrl_dev_get_drvdata(pctldev);
 	enum pin_config_param param;
 	int i, ret;
-	u32 arg;
+	u16 arg;
 
 	if (chv_pad_locked(pctrl, pin))
 		return -EBUSY;
