+++ drivers/pinctrl/pinctrl-rockchip.c
@@ -1386,7 +1386,7 @@ static int rockchip_pinconf_set(struct pinctrl_dev *pctldev, unsigned int pin,
 	struct rockchip_pinctrl *info = pinctrl_dev_get_drvdata(pctldev);
 	struct rockchip_pin_bank *bank = pin_to_bank(info, pin);
 	enum pin_config_param param;
-	u32 arg;
+	u16 arg;
 	int i;
 	int rc;
